+++ schema/changelog-3.16.xml
@@ -20,5 +20,11 @@
         <column name="poilayer" type="VARCHAR(512)" />
     </addColumn>
 
+    <addColumn tableName="notifications">
+      <column name="calendarid" type="INT" />
+    </addColumn>
+
+    <addForeignKeyConstraint baseColumnNames="calendarid" baseTableName="notifications" constraintName="fk_notification_calendar_calendarid" onDelete="SET NULL" onUpdate="RESTRICT" referencedColumnNames="id" referencedTableName="calendars"/>
+
   </changeSet>
 </databaseChangeLog>
